salary check in test + moved filters into comp
